changing how errors happen on github actions
