+++ .github/workflows/FloxBx.yml
@@ -119,9 +119,12 @@ jobs:
             ${{ runner.os }}-build-
             ${{ runner.os }}-
         if: startsWith(matrix.xcode,'/Applications/Xcode_14')
-      - name: Set Lint Strictness
-        run: echo "LINT_MODE=STRICT" >> $GITHUB_ENV
-        if: ${{ github.event_name == 'pull_request' }} && ${{ github.base_ref == 'main' }}
+      # - name: Set Lint Strictness
+      #   run: echo "LINT_MODE=STRICT" >> $GITHUB_ENV
+      #   if: ${{ github.head_ref == 'main' }}
+      # - name: Set Throw Errors on Lint
+      #   run: echo "THROW_ERRORS=YES" >> $GITHUB_ENV
+      #   if: ${{ github.head_ref == 'main' }}
       - name: Set Xcode Name
         run: echo "XCODE_NAME=$(basename -- ${{ matrix.xcode }} | sed 's/\.[^.]*$//' | cut -d'_' -f2)" >> $GITHUB_ENV
       - name: Setup Xcode
