+++ Packages/FloxBxKit/Sources/FloxBxModels/Content/CreateGroupSessionResponseContent.swift
@@ -1,0 +1,8 @@
+import Foundation
+public struct CreateGroupSessionResponseContent: Codable {
+  public let id: UUID
+
+  public init(id: UUID) {
+    self.id = id
+  }
+}
