+++ Packages/FloxBxKit/Sources/FloxBxModels/Content/UserSubscriptionRequestContent.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct UserSubscriptionRequestContent: Codable {
+  public let tags: [String]
+}
