+++ Packages/FloxBxKit/Sources/FloxBxModels/Notifications/Notifiable.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public protocol Notifiable: Codable {
+  associatedtype PayloadType: Codable
+  var title: String { get }
+  var deviceToken: Data { get }
+  var topic: String { get }
+  var payload: PayloadType { get }
+}
