+++ Packages/FloxBxKit/Sources/FloxBxModels/Notifications/NotificationContent.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol NotificationContent {
+  var title: String { get }
+}
