+++ Packages/FloxBxKit/scripts/lint.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 
+SomeErrorHandler () {
+	(( errcount++ ))       # or (( errcount += $? ))
+}
+
 if [ -z "$SRCROOT" ]; then
 	SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 	PACKAGE_DIR="${SCRIPT_DIR}/.."
@@ -11,6 +15,8 @@ if [ -z "$GITHUB_ACTION" ]; then
 	MINT_CMD="/opt/homebrew/bin/mint"
 else
 	MINT_CMD="mint"
+	trap SomeErrorHandler ERR
+	LINT_MODE="STRICT"
 fi
 
 export MINT_PATH="$PACKAGE_DIR/.mint"
@@ -46,3 +52,8 @@ $MINT_RUN swiftformat --lint $SWIFTFORMAT_OPTIONS .
 $MINT_RUN swiftlint lint $SWIFTLINT_OPTIONS
 
 popd
+
+if  [ $errcount > 0 ]; then
+	echo "Too many errors"
+	exit $errcount
+fi
